further separation of the code into logical locations
