+++ pkg/storage/loaders/data_loaders.go
@@ -1,0 +1,25 @@
+package loaders
+
+import "github.com/cmsgov/mint-app/pkg/storage"
+
+// DataLoaders wrap your data loaders to inject via middleware
+type DataLoaders struct {
+	BasicsLoader            *WrappedDataLoader
+	OperationalNeedLoader   *WrappedDataLoader
+	OperationSolutionLoader *WrappedDataLoader
+	DataReader              *DataReader
+}
+
+// NewDataLoaders instantiates data loaders for the middleware
+func NewDataLoaders(store *storage.Store) *DataLoaders {
+	loaders := &DataLoaders{
+		DataReader: &DataReader{
+			Store: store,
+		},
+	}
+	loaders.BasicsLoader = newWrappedDataLoader(loaders.GetPlanBasicsByModelPlanID)
+	loaders.OperationalNeedLoader = newWrappedDataLoader(loaders.GetOperationalNeedsByModelPlanID)
+	loaders.OperationSolutionLoader = newWrappedDataLoader(loaders.GetOperationalSolutionAndPossibleCollectionByOperationalNeedID)
+
+	return loaders
+}
